Fix format-string warnings for time_t
The `time_t` type is defined as `long` on 64-bit Unix platforms, which
is wide enough to deal with the year-2083 problem. To make sure the code
is standards-compliant on other platforms, we cast the `time_t` values
to `long`.

+++ src/bgp/bgp.c
@@ -656,9 +656,9 @@ void skinny_bgp_daemon_online()
 		bgp_peer_close(&peers[peers_check_idx], FUNC_TYPE_BGP, FALSE, FALSE, FALSE, FALSE, NULL);
 	      }
 	      else {
-		Log(LOG_WARNING, "WARN ( %s/%s ): [%s] Refusing new connection from existing peer (residual holdtime: %u).\n",
+		Log(LOG_WARNING, "WARN ( %s/%s ): [%s] Refusing new connection from existing peer (residual holdtime: %ld).\n",
 			config.name, bgp_misc_db->log_str, bgp_peer_str,
-			(peers[peers_check_idx].ht - (now - peers[peers_check_idx].last_keepalive)));
+			(peers[peers_check_idx].ht - ((long)now - peers[peers_check_idx].last_keepalive)));
 		FD_CLR(peer->fd, &bkp_read_descs);
 		bgp_peer_close(peer, FUNC_TYPE_BGP, FALSE, FALSE, FALSE, FALSE, NULL);
 		goto read_data;
